Add files via upload
In this code, the datetime module is imported to enable working with dates. The input function is used to prompt the user to enter a date in dd/mm/yyyy format.

The map function is used to split the input string into three integers representing the day, month, and year. These values are then used to create a datetime.date object using the datetime.date function.

The strftime method is used to format the resulting date object as a string, with the %A format specifier representing the full name of the day of the week.

The print function is used to display the day of the week corresponding to the entered date.